integration_tests/requirements.txt
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0
aiohttp==3.9.3
requests==2.31.0
python-dotenv==1.0.1
fastapi==0.109.2
httpx==0.24.0
pydantic>=2.0.0 